extract getSalesByTitle controller
